Fixes random extension disconnects
These apparently are caused by the browser's idle timeout for background scripts, which is enforced more strictly in Manifest v3.
There is no manifest option to make the background script persistent,
and using a keepalive logic is fallible because the browser will
restrict JS execution time for idle pages.

The solution is to automatically reconnect the
contentscript<->background connection on disconnect, and reseting all
the appropriate callbacks.